backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
firebase-admin==6.5.0
